move refactored agent into task
